Refactor main file and create departments file
Copy paste const departments in own 'departments' file.
