+++ scripts/main.js
@@ -1,74 +1,33 @@
-const departments = {
-    marketing: {
-        description: "Marketing omvat verschillende strategieën en activiteiten gericht op het promoten van producten, diensten of merken bij het doelpubliek. Het omvat het begrijpen van de behoeften en voorkeuren van consumenten, het uitvoeren van marktonderzoek, het ontwikkelen van marketingplannen en het implementeren van tactieken om potentiële klanten te bereiken en te betrekken. Marketingprofessionals analyseren markttrends, identificeren doelmarkten en creëren overtuigende boodschappen en communicatiestrategieën om merkbekendheid op te bouwen, klantenwerving te stimuleren en klantloyaliteit te bevorderen",
-        numberOfEmployees: 45,
-        jobs: [
-            {
-                title: "Marketingmanager",
-                description: "De marketingmanager is verantwoordelijk voor het ontwikkelen en implementeren van marketingstrategieën om de producten of diensten van een bedrijf te promoten. Ze houden toezicht op marktonderzoek, reclamecampagnes en initiatieven voor klantenwerving. Ze analyseren markttrends, identificeren doelgroepen en werken samen met interdisciplinaire teams om marketingdoelen te bereiken en bedrijfsgroei te stimuleren."
-            },
-            {
-                title: "Digital Marketing Specialist",
-                description: "Een Digital Marketing Specialist richt zich op online marketingactiviteiten om de zichtbaarheid van het merk, websiteverkeer en klantbetrokkenheid te vergroten. Ze maken gebruik van verschillende digitale kanalen zoals sociale media, e-mailmarketing, zoekmachineoptimalisatie (SEO) en betaalde advertenties (PPC). Ze monitoren de prestaties van campagnes, analyseren gegevens en optimaliseren strategieën om online aanwezigheid te maximaliseren en leads of conversies te genereren."
-            },
-            {
-                title: "Contentmarketeer",
-                description: "Een contentmarketeer is verantwoordelijk voor het ontwikkelen en uitvoeren van contentstrategieën die aansluiten bij het merk en de marketingdoelstellingen van een bedrijf. Ze houden toezicht op contentcreatie, waaronder blogs, artikelen, video's, infographics en berichten op sociale media. Ze werken samen met schrijvers, ontwerpers en andere belanghebbenden om hoogwaardige content te produceren die de doelgroep aanspreekt en merkbekendheid, thought leadership en klantenwerving stimuleert."
-            },
-            {
-                title: "Branding Agent",
-                description: "Een Branding Agent is verantwoordelijk voor het vaststellen en handhaven van een consistente merkafbeelding en identiteit op verschillende marketingkanalen. Ze ontwikkelen richtlijnen voor het merk, positionering en berichtenstrategieën. Ze houden toezicht op productlanceringen, reclamecampagnes en inspanningen op het gebied van merkcommunicatie. Ze volgen de prestaties van het merk, voeren marktonderzoek uit en doen aanbevelingen om de merkperceptie en merkwaarde te verbeteren."
-            },
-            {
-                title: "Marktonderzoeksanalist",
-                description: "Een marktonderzoeksanalist voert diepgaand onderzoek uit om gegevens te verzamelen en te analyseren over markttrends, consumentenvoorkeuren en concurrentieactiviteiten. Ze ontwerpen en voeren enquêtes, focusgroepen en interviews uit om relevante informatie te verzamelen. Ze interpreteren gegevens, genereren inzichten en bereiden rapporten voor met bruikbare aanbevelingen. Ze helpen bedrijven geïnformeerde beslissingen te nemen over productontwikkeling, doelmarkten, prijsstrategieën en marketingcampagnes."
-            },
-        ],
-    },
-    sales: {
-        description: "Sales omvat activiteiten en strategieën die gericht zijn op het genereren van omzet door het verkopen van producten, diensten of oplossingen aan klanten. Het draait om het opbouwen van relaties, het identificeren van klantbehoeften en het overtuigend communiceren van de waarde en voordelen van een product of dienst aan potentiële kopers. Verkoopprofessionals spelen een cruciale rol bij het stimuleren van bedrijfsgroei en het behalen van omzetdoelstellingen. Ze zoeken naar en kwalificeren leads, hebben directe interacties met klanten, onderhandelen over contracten en sluiten deals.",
-        numberOfEmployees: 23,
-        jobs: [
-            {
-                title: "Sales Representative",
-                description: "Een Sales Representative is verantwoordelijk voor actief zoeken naar en genereren van nieuwe leads, contact opnemen met potentiële klanten en producten of diensten promoten en verkopen. Ze bouwen relaties op met klanten, begrijpen hun behoeften en bieden oplossingen die aan hun eisen voldoen. Verkoopvertegenwoordigers geven vaak productdemonstraties, onderhandelen over contracten en sluiten verkoopdeals."
-            },
-            {
-                title: "Verkoopmanager",
-                description: "Een verkoopmanager geeft leiding aan een team van verkoopprofessionals en is verantwoordelijk voor het stellen van verkoopdoelen, het ontwikkelen van verkoopstrategieën en het bewaken van de prestaties van het team. Ze bieden begeleiding, coaching en training aan het verkoopteam om het behalen van verkoopdoelstellingen te waarborgen. Verkoopmanagers werken ook samen met andere afdelingen, zoals marketing en productontwikkeling, om verkoopinspanningen af te stemmen op algemene bedrijfsdoelen."
-            },
-            {
-                title: "Accountmanager",
-                description: "Een accountmanager richt zich op het beheren en uitbreiden van relaties met bestaande klanten. Ze werken nauw samen met klanten om hun zakelijke behoeften te begrijpen, oplossingen voor te stellen en aanvullende producten of diensten te verkopen. Accountmanagers onderhouden regelmatig contact met klanten, pakken eventuele zorgen of problemen aan en streven naar klanttevredenheid en behoud."
-            },
-            {
-                title: "Business Development Manager",
-                description: "Een business development manager identificeert en volgt kansen voor bedrijfsgroei en uitbreiding. Ze onderzoeken en analyseren markttrends, identificeren potentiële partnerschappen of allianties en ontwikkelen strategieën om nieuwe markten of klantsegmenten te betreden. Business development managers nemen ook deel aan netwerkactiviteiten, bezoeken brancheevenementen en onderhandelen over strategische partnerschappen of contracten om de omzetgroei te stimuleren."
-            }
-        ]
-    },
-    "customer-service": {
-        description: "Klantenservice richt zich op het bieden van uitzonderlijke ondersteuning en assistentie aan klanten vóór, tijdens en na hun interacties met een bedrijf. Het draait om het opbouwen van positieve relaties met klanten en ervoor zorgen dat ze tevreden zijn door hun vragen, zorgen en behoeften aan te pakken. Klantenserviceprofessionals fungeren als de vertegenwoordigers van een bedrijf, als het eerste aanspreekpunt voor klanten. Ze maken gebruik van verschillende communicatiekanalen, zoals telefoongesprekken, e-mails, live chats of persoonlijke interacties, om klanten op een vriendelijke, efficiënte en empathische manier te helpen.",
-        numberOfEmployees: 32,
-        jobs: [
-            {
-                title: "Klantenservicemedewerker",
-                description: "Een klantenservicemedewerker fungeert als het primaire contactpunt voor klanten. Ze behandelen vragen, verstrekken product- of service-informatie, gaan in op klachten of zorgen van klanten en zorgen voor tijdige oplossingen. Klantenservicemedewerkers communiceren via verschillende kanalen zoals telefoongesprekken, e-mails of live chats en streven ernaar uitstekende klantervaringen te bieden."
-            },
-            {
-                title: "Klantenondersteuningsspecialist",
-                description: "Een klantenondersteuningsspecialist biedt technische ondersteuning en probleemoplossing aan klanten die problemen ondervinden met een product of dienst. Ze hebben een grondig begrip van het product of de dienst en begeleiden klanten effectief bij het oplossen van problemen. Klantenondersteuningsspecialisten kunnen ook begeleiding geven over het gebruik, de installatie of upgrades van het product."
-            },
-            {
-                title: "Client Success Manager",
-                description: "Een accountmanager klantsucces richt zich op het opbouwen en onderhouden van sterke relaties met belangrijke klanten. Ze werken nauw samen met klanten om hun doelen te begrijpen, aan hun behoeften te voldoen en ervoor te zorgen dat ze waarde halen uit de geboden producten of diensten. Accountmanagers klantsucces engageren zich proactief met klanten, bieden begeleiding en identificeren kansen voor upselling of cross-selling."
-            },
-            {
-                title: "Customer Experience Manager",
-                description: "A Customer Experience Manager is verantwoordelijk voor het creëren en implementeren van strategieën om de algehele klantervaring te verbeteren. Ze analyseren klantinteracties, identificeren pijnpunten en ontwikkelen initiatieven om de klanttevredenheid en -loyaliteit te vergroten. Klantbelevingsmanagers werken samen met verschillende afdelingen om de klantgerichtheid van het bedrijf te verbeteren en klantgerichte processen en beleid te ontwikkelen."
-            }
-        ],
-    }
-}
+import departments from "./departments.js";
+import {
+    getDepartmentInformation,
+    getDesiredRoleInformation,
+    getRoleDescription
+} from "./departmentFunctions.js";
+import {pushPromptedInformation} from "./uiHelperFunctions.js";
+
 
 console.log(departments);
+
+/* (Opdracht 1 - Gegevens aanspreken) */
+
+const salesManagerDescription = departments.sales.jobs.find(job => job.title === "Verkoopmanager").description;
+
+console.log(`De afdeling Sales heeft ${departments.sales.numberOfEmployees} medewerkers.`);
+console.log(`Marketing is een leuke afdeling om te werken. ${departments.marketing.description}`);
+console.log(`De afdeling Customer Service heeft ${departments["customer-service"].numberOfEmployees} medewerkers.`);
+console.log(`Sales is een uitdagende afdeling om te werken als Verkoopmanager. ${salesManagerDescription}` );
+
+/*
+    (Opdracht 2 - prompten en beslissen)
+    (Opdracht 3 - een nieuwe prompt)
+    (Opdracht 4 - samenvoegen)
+    (Opdracht 5 - bonus)
+*/
+
+const deptInput = prompt("Over welke afdeling wil je meer informatie? Kies uit: [marketing / sales / customer-service].").toLowerCase();
+const deptRoleInput = getDesiredRoleInformation(deptInput);
+
+getDepartmentInformation(deptInput);
+getRoleDescription(deptInput, deptRoleInput);
+pushPromptedInformation(deptInput, deptRoleInput); /* (Opdracht 5 - bonus)*/
